refactor: 채팅방 상세 조회 로직을 별도로 분리
